pyproject.toml
[tool.poetry]
name = "supergene"
version = "0.1.0"
description = "A library to automate the scraping, editing, formatting of the webnovel `Super Gene` into ePub ebooks."
authors = ["Max Ludden <dev@maxludden.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
mongoengine = "^0.24.2"
tqdm = "^4.64.1"
num2words = "^0.5.12"
requests = "^2.28.1"
bs4 = "^0.0.1"
selenium = "^4.4.3"
sh = "^1.14.3"
loguru = "^0.6.0"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"
pyaml = "^21.10.1"
beanie = "^1.11.9"
fastapi = "^0.82.0"
uvicorn = "^0.18.3"
python-cfonts = "^1.5.2"
rich = "^12.6.0"
ipython = "^8.5.0"
ipywidgets = "^8.0.2"
typer = {extras = ["all"], version = "^0.6.1"}
maxcolor = "^0.4.0"

[tool.poetry.scripts]
supergene = "supergene.__main__:main"

[tool.poetry.group.dev.dependencies]
mongo-types = "^0.15.1"
ujson = "^5.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
